Refactor entity package to model
