+++ src/application/csv/Csv.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import application.entity.Linha;
+import application.model.Linha;
 import application.util.LoggerUtil;
 
 /**
